app/requirement.txt
aiohttp==3.7.4.post0
alembic==1.5.6
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.7.0
astroid==2.5.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chardet2==2.0.3
click==7.1.2
colorama==0.4.4
colorhash==1.0.3
config==0.5.0.post0
configparser==5.0.2
crontab==0.22.9
cryptography==3.4.7
decorator==5.0.7
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
flasgger==0.9.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-FlatPages==0.7.2
Flask-FlatPages-Pandoc==0.2
Flask-JSON==0.3.4
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.7.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-Testing==0.8.1
Frozen-Flask==0.15
future==0.18.2
gcloud==0.18.3
geocoder==1.38.1
gevent==21.1.2
google-api-core==1.26.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grequests==0.6.0
html5lib==1.1
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.3
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonschema==3.2.0
jsonschema2md==0.1.1
lazy-object-proxy==1.5.2
logger==1.4
lxml==4.6.2
Mako==1.1.4
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==1.1.1
marshmallow==3.11.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
Password-Validation==0.1.1
pbr==5.5.1
pdoc3==0.9.2
Pillow==8.2.0
pip==21.0.1
pipenv==2020.11.15
priority==1.3.0
protobuf==3.15.6
psutil==5.8.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pycurl==7.43.0.5
PyJWT==2.0.1
pylint==2.7.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-editor==1.0.4
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
pytz==2021.1
PyYAML==5.4.1
ratelim==0.1.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
schedule==1.0.0
scipy==1.6.3
setuptools==52.0.0.post20210125
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
stevedore==3.3.0
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.3
urlopen==1.0.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper.basic==0.2.2
webencodings==0.5.1
Werkzeug==0.16.1
wheel==0.36.2
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
